+++ pkg/deployment/reconcile/action_helper.go
@@ -38,6 +38,12 @@ func NewTimeoutFetcher(t time.Duration) TimeoutFetcher {
 	}
 }
 
+type actionEmpty struct {
+	actionImpl
+	actionEmptyStart
+	actionEmptyCheckProgress
+}
+
 type actionEmptyCheckProgress struct {
 }
 
@@ -102,7 +108,7 @@ func (a actionImpl) Timeout(deploymentSpec api.DeploymentSpec) time.Duration {
 	return a.timeout(deploymentSpec)
 }
 
-// Return the MemberID used / created in this action
+// MemberID returns the member ID used / created in the current action.
 func (a actionImpl) MemberID() string {
 	return *a.memberIDRef
 }