+++ pkg/version/version.community.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2020 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,30 +17,11 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Adam Janikowski
-// Author Tomasz Mielech
-//
 
-package features
+// +build community
 
-import (
-	"github.com/arangodb/kube-arangodb/pkg/util"
+package version
+
+var (
+	edition = CommunityEdition
 )
-
-func init() {
-	registerFeature(metricsExporter)
-}
-
-var metricsExporter = &feature{
-	name:               "metrics-exporter",
-	description:        "Define if internal metrics-exporter should be used",
-	version:            "3.6.0",
-	enterpriseRequired: false,
-	enabledByDefault:   true,
-	deprecated:         "It is always set to True",
-	constValue:         util.NewBool(true),
-}
-
-func MetricsExporter() Feature {
-	return metricsExporter
-}