+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
+- Replace `beta.kubernetes.io/arch` Pod label with `kubernetes.io/arch` using Silent Rotation
+- Add "Short Names" feature
+- Switch ArangoDB Image Discovery process from Headless Service to Pod IP
+
+## [1.2.3](https://github.com/arangodb/kube-arangodb/tree/1.2.3) (2021-09-24)
 - Update UBI Image to 8.4
-- Fix ArangoSync Liveness Prove
+- Fix ArangoSync Liveness Probe
 - Allow runtime update of Sidecar images
 - Allow Agent recreation with preserved IDs
+- The internal metrics exporter can not be disabled
+- Changing the topics' log level without restarting the container.
+  When the topic is removed from the argument list then it will not 
+  be turned off in the ArangoDB automatically.
+- Allow to customize SchedulerName inside Member Pod
+- Add Enterprise Edition support
 
 ## [1.2.2](https://github.com/arangodb/kube-arangodb/tree/1.2.2) (2021-09-09)
 - Update 'github.com/arangodb/arangosync-client' dependency to v0.7.0